+++ src/bitmap_smooth.c
@@ -20,7 +20,7 @@ extern _iobuf *__iob_func(void);
 extern int fprintf(_iobuf *stream, const char *format, ...);
 extern int fflush(_iobuf *stream);
 
-extern float floor(float x);
+extern double floor(double x);
 extern void bitmap_3d_smooth(bitmap_data *bitmap, int16_t filter_size, const int16_t *filter_coefficients);
 extern void bitmap_2d_smooth(bitmap_data *bitmap, int16_t filter_size, const int16_t *filter_coefficients);
